requirements.txt
torchvision==0.20.1
torch==2.5.1
streamlit==1.39.0
pandas==2.2.2
scikit-learn==1.5.1
pillow==10.4.0